app_listen/requirements.txt
paho-mqtt
httpx
asyncio
mysql-connector-python